feat: test for feature switches

+++ src/main/java/de/caritas/cob/userservice/api/deleteworkflow/scheduler/DeleteInactiveSessionsAndUserScheduler.java
@@ -3,6 +3,7 @@ package de.caritas.cob.userservice.api.deleteworkflow.scheduler;
 import de.caritas.cob.userservice.api.deleteworkflow.service.DeleteInactiveSessionsAndUserService;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -12,8 +13,16 @@ public class DeleteInactiveSessionsAndUserScheduler {
 
   private final @NonNull DeleteInactiveSessionsAndUserService deleteInactiveSessionsAndUserService;
 
+  @Value("${session.inactive.deleteWorkflow.enabled}")
+  private boolean sessionInactiveDeleteWorkflowEnabled;
+
   @Scheduled(cron = "${session.inactive.deleteWorkflow.cron}")
   public void performDeletionWorkflow() {
+
+    if (!sessionInactiveDeleteWorkflowEnabled) {
+      return;
+    }
+
     this.deleteInactiveSessionsAndUserService.deleteInactiveSessionsAndUsers();
   }
